+++ .gitignore
@@ -1,0 +1,2 @@
+simulation.mp4
+__pycache__
