+++ Tests.py
@@ -1,0 +1,6 @@
+import unittest
+
+from tests.NodeTests import *
+
+if __name__=='__main__':
+    unittest.main(verbosity=2)
